Added guard rails when running after a partial cleardown
The deployment process would previously be blocked if a previous cleardown had not deleted the default 'tenancy administrator' service principal - as the associated credential would no longer be available (if the instance keyvault had been deleted).

Now we simply reset the credential if we don't have it in keyvault.

NOTE: This assumes no-one else has a reason to use this service principal.