Work around recently deleted container test problems
Azure Storage gives you a 409 if you try to create a container with the same name as one you deleted recently, and this causes us problems if we try to clean up after creating tenants with well-known ids. So this change only deletes the blobs for such tenants when cleaning up.